Refactor model.user package to model.account
