+++ src/main/java/com/example/echo_api/persistence/repository/UserRepository.java
@@ -5,7 +5,7 @@ import java.util.UUID;
 import org.springframework.data.repository.ListCrudRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.echo_api.persistence.model.user.User;
+import com.example.echo_api.persistence.model.account.User;
 
 import java.util.Optional;
