+++ src/main/java/com/example/echo_api/security/AuthenticationConfig.java
@@ -10,7 +10,7 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
-import com.example.echo_api.persistence.model.user.SecurityUser;
+import com.example.echo_api.persistence.model.account.SecurityUser;
 import com.example.echo_api.persistence.repository.UserRepository;
 
 import lombok.RequiredArgsConstructor;
