+++ src/main/java/com/example/echo_api/service/profile/ProfileService.java
@@ -3,8 +3,8 @@ package com.example.echo_api.service.profile;
 import com.example.echo_api.exception.custom.username.UsernameException;
 import com.example.echo_api.persistence.dto.request.profile.UpdateProfileDTO;
 import com.example.echo_api.persistence.dto.response.profile.ProfileDTO;
+import com.example.echo_api.persistence.model.account.User;
 import com.example.echo_api.persistence.model.profile.Profile;
-import com.example.echo_api.persistence.model.user.User;
 
 public interface ProfileService {
